implement TradeDataFrame to read archive
