+++ rbot_lib/src/db/archive.rs
@@ -59,6 +59,26 @@ pub struct TradeArchive {
     end_time: MicroSec,
 }
 
+impl Clone for TradeArchive {
+    fn clone(&self) -> Self {
+        let mut archive = Self {
+            config: self.config.clone(),
+            production: self.production.clone(),
+            last_archive_check_time: self.last_archive_check_time.clone(),
+            latest_archive_date: self.latest_archive_date.clone(),
+            start_time: self.start_time.clone(),
+            end_time: self.end_time.clone(),
+        };
+
+        let r = archive.analyze();
+        if r.is_err() {
+            log::warn!("trade archive analyze error");
+        }
+
+        archive
+    }
+}
+
 impl TradeArchive {
     pub fn new(config: &MarketConfig, production: bool) -> Self {
         let mut my = Self {
@@ -75,18 +95,34 @@ impl TradeArchive {
             log::debug!("Archive analyze error {:?}", r);
         }
 
-        log::debug!("S: {} ({})", my.start_time(), time_string(my.start_time()));
-        log::debug!("E: {} ({})", my.end_time(), time_string(my.end_time()));
+        log::debug!(
+            "S: {:?} ({:?})",
+            my.start_time(),
+            time_string(my.start_time().unwrap_or_default())
+        );
+        log::debug!(
+            "E: {:?} ({:?})",
+            my.end_time(),
+            time_string(my.end_time().unwrap_or_default())
+        );
 
         return my;
     }
 
-    pub fn start_time(&self) -> MicroSec {
-        self.start_time
+    pub fn start_time(&self) -> anyhow::Result<MicroSec> {
+        if self.start_time == 0 {
+            return Err(anyhow!("no data in archive"));
+        }
+
+        Ok(self.start_time)
     }
 
-    pub fn end_time(&self) -> MicroSec {
-        self.end_time
+    pub fn end_time(&self) -> anyhow::Result<MicroSec> {
+        if self.end_time == 0 {
+            return Err(anyhow!("no data in archive"));
+        }
+
+        Ok(self.end_time)
     }
 
     // check if the log data is already stored in local.
@@ -217,20 +253,20 @@ impl TradeArchive {
     pub fn load_cache_df(&self, date: MicroSec) -> anyhow::Result<DataFrame> {
         let date = FLOOR_DAY(date);
 
-        if date < self.start_time() {
+        if date < self.start_time()? {
             log::warn!(
                 "Not found in archive[too early] query={:?} < start_time{:?}",
                 date_string(date),
-                date_string(self.start_time())
+                date_string(self.start_time()?)
             );
             return Ok(Self::make_empty_cachedf());
         }
 
-        if self.end_time() <= date {
+        if self.end_time()? <= date {
             log::warn!(
                 "Not found in archive[too new] query={:?} >= end_time{:?}",
                 date_string(date),
-                date_string(self.end_time())
+                date_string(self.end_time()?)
             );
             return Ok(Self::make_empty_cachedf());
         }
@@ -323,14 +359,14 @@ impl TradeArchive {
         start_time: MicroSec,
         end_time: MicroSec,
     ) -> anyhow::Result<Vec<MicroSec>> {
-        let start_time = if start_time == 0 || start_time < self.start_time() {
-            self.start_time()
+        let start_time = if start_time == 0 || start_time < self.start_time()? {
+            self.start_time()?
         } else {
             FLOOR_DAY(start_time)
         };
 
-        let end_time = if end_time == 0 || self.end_time() <= end_time {
-            FLOOR_DAY(self.end_time() - 1)
+        let end_time = if end_time == 0 || self.end_time()? <= end_time {
+            FLOOR_DAY(self.end_time()? - 1)
         } else {
             FLOOR_DAY(end_time)
         };
