+++ rbot_lib/src/db/mod.rs
@@ -4,10 +4,12 @@ pub mod sqlite;
 pub mod df;
 pub mod fs;
 pub mod archive;
+pub mod tradedf;
 
 pub use sqlite::*;
 pub use df::*;
 pub use fs::*;
 pub use archive::*;
+pub use tradedf::*;
